scrimish_game_mobile_application
Scrimish is a fast moving card game that pits player against player in an epic 10 minute battle of strategy, memory, and misdirection. Every game is different, and you'll want to play over and over again to try new ways of outsmarting your opponent. The game is expandable to large team or free-for-all games.